Move verification to internal package
